Image_enhancement/requirements.txt
numpy==1.24.2
pandas==2.0.0
Pillow==9.0.1
Flask==2.3.3
wget==3.2
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
pika==1.3.1
toml==0.10.2
requests==2.27.1
my_module==1.6.2
easydev==0.12.0
colormap==1.0.4
colorgram.py==1.2.0
extcolors==1.0.0
mysql-connector-python==8.1.0
protobuf==4.21.12
scikit-learn==1.3.0
matplotlib==3.7.2
pymongo==4.4.0
gunicorn==21.2.0
imagehash==4.3.1
phpserialize==1.3
validators==0.22.0
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
scikit-image==0.22.0
psutil==5.9.8
